Moved test files to tests package
